- Fixed: The sky scrolling calculations caused an integer multiplication overflow. - Fixed: Voodoo dolls should not start ENTER scripts. - Fixed: ActorDamage must not parse the enclosing parentheses as part of the   expression. That will produce an error if a flag set or clear command immediately   follows. - Fixed: P_DamageMobj ignored MF2_NODMGTHRUST if the damaging object had no   owner. - Added a 'font' parameter to A_Print. - Changed A_CustomMeleeAttack to take one damage parameter only. Since expressions   can be used this value is not used as a factor for a random value but as   a direct damage value instead. - Fixed: AActor::SetState must check whether a called action function resulted   in the actor's destruction. A_Jump constructs to a 0-length terminating   state will hang if this isn't checked.
SVN r329 (trunk)

+++ src/r_anim.cpp
@@ -802,7 +802,7 @@ void R_UpdateAnimations (DWORD mstime)
 	}
 
 	// Scroll the sky
-	double ms = mstime * FRACUNIT;
+	double ms = (double)mstime * FRACUNIT;
 	sky1pos = fixed_t(fmod (ms * level.skyspeed1, double(TexMan[sky1texture]->GetWidth() << FRACBITS)));
 	sky2pos = fixed_t(fmod (ms * level.skyspeed2, double(TexMan[sky2texture]->GetWidth() << FRACBITS)));
 }
